client_requirements.txt
boto3
fire
jinja2
loguru
pynacl
python-dotenv
requests
